+++ src/Products/urban/Corporation.py
@@ -158,7 +158,7 @@ Corporation_schema = BaseSchema.copy() + \
 Corporation_schema['society'].widget.visible = False
 Corporation_schema['representedBySociety'].widget.visible = False
 Corporation_schema['representedBy'].widget.visible = False
-Corporation_schema['nationalRegister'].widget.visible = False
+# Corporation_schema['nationalRegister'].widget.visible = False
 #Corporation_schema['personTitle'].widget.visible = False
 ##/code-section after-schema
 
@@ -224,7 +224,8 @@ def finalizeSchema(schema, folderish=False, moveDiscussion=True):
     schema.moveField('legalForm', after='denomination')
     schema.moveField('tvaNumber', after='fax')
     schema.moveField('bceNumber', after='tvaNumber')
-    schema.moveField('personTitle', after='bceNumber')
+    schema.moveField('nationalRegister', after='bceNumber')
+    schema.moveField('personTitle', after='nationalRegister')
     schema.moveField('personRole', after='personTitle')
     schema.moveField('name1', after='personRole')
     schema.moveField('name2', after='name1')